+++ conf/worker.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-exec celery -A miracle.worker.app:celery_app worker -Ofair --no-execv --without-mingle --without-gossip
+exec celery -A miracle.worker.app:celery_app worker -P gevent -Ofair --no-execv --without-mingle --without-gossip