+++ miracle/conftest.py
@@ -53,12 +53,6 @@ def teardown_db(engine):
 
 @pytest.fixture(scope='session', autouse=True)
 def package():
-    # Apply gevent monkey patches as early as possible during tests.
-    from gevent.monkey import patch_all
-    patch_all()
-    from psycogreen.gevent import patch_psycopg
-    patch_psycopg()
-
     # Enable all warnings in test mode.
     warnings.resetwarnings()
     warnings.simplefilter('default')