+++ miracle/worker/app.py
@@ -38,7 +38,7 @@ def init_worker(celery_app,
                 _bloom_domain=None, _cache=None,
                 _crypto=None, _db=None, _raven=None, _stats=None):
     configure_logging()
-    raven = create_raven(transport='threaded', _raven=_raven)
+    raven = create_raven(transport='gevent', _raven=_raven)
 
     try:
         celery_app.bloom_domain = create_bloom_domain(_bloom=_bloom_domain)