+++ miracle/worker/settings.py
@@ -1,4 +1,3 @@
-import multiprocessing
 import os
 
 testing = bool('TESTING' in os.environ)
@@ -23,7 +22,7 @@ CELERY_IMPORTS = [
 ]
 
 # Optimization
-CELERYD_CONCURRENCY = multiprocessing.cpu_count()
+CELERYD_CONCURRENCY = 20
 CELERYD_MAX_TASKS_PER_CHILD = 100000
 CELERYD_PREFETCH_MULTIPLIER = 8
 CELERY_DISABLE_RATE_LIMITS = True
@@ -33,6 +32,5 @@ CELERY_RESULT_SERIALIZER = 'json'
 CELERY_TASK_SERIALIZER = 'json'
 
 # cleanup
-del multiprocessing
 del os
 del testing